Switch install logic to app to prevent memory leak
