+++ app/containers/App.js
@@ -1,6 +1,9 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import _ from 'lodash';
+import { ipcRenderer } from 'electron';
+import { exec } from 'child_process';
+import DecompressZip from 'decompress-zip';
 import type { Children } from 'react';
 
 import SideBar from '../components/SideBar/SideBar';
@@ -8,6 +11,7 @@ import TopBar from '../components/TopBar';
 import Auth from '../components/Auth/Auth';
 
 import { removeUser, updateProfilePic } from '../actions/user';
+import { startInstall, finishInstall } from '../actions/game';
 
 const mapStateToProps = ({ user }) => (
   {
@@ -20,8 +24,12 @@ const mapDispatchToProps = dispatch => ({
     localStorage.removeItem('token');
     dispatch(removeUser());
   },
-  updateUserPic: profilePic => dispatch(updateProfilePic(profilePic))
+  updateUserPic: profilePic => dispatch(updateProfilePic(profilePic)),
+  startInstalling: () => dispatch(startInstall()),
+  completeInstall: () => dispatch(finishInstall())
 });
+
+let isInstalling = false;
 
 class App extends Component {
   props: {
@@ -29,7 +37,58 @@ class App extends Component {
     history: {},
     user: {},
     logout: Function, //eslint-disable-line
-    updateUserPic: Function //eslint-disable-line
+    updateUserPic: Function, //eslint-disable-line
+    startInstalling: Function, //eslint-disable-line
+    completeInstall: Function //eslint-disable-line
+  };
+
+  componentDidMount() {
+    const { startInstalling } = this.props;
+    ipcRenderer.on('download-finish', (event, args) => {
+      const { savePath, url } = args;
+
+      const brokenUrl = url.split('/');
+      const filename = brokenUrl[brokenUrl.length - 1];
+      const unzipTo = savePath.substring(0, savePath.length - filename.length);
+
+      if (isInstalling) return;
+      isInstalling = true;
+      startInstalling();
+
+      if (process.platform === 'darwin') this.unzipMac(savePath, unzipTo);
+      else this.unzipWin(savePath, unzipTo);
+    });
+  }
+
+  unzipMac = (savePath, unzipTo) => {
+    const { completeInstall } = this.props;
+    exec(`unzip ${savePath} -d ${unzipTo}`, (error) => {
+      if (error) { throw error; }
+      completeInstall();
+      isInstalling = false;
+
+      // Delete .zip after unzipping
+      exec(`rm -rf ${savePath}`, (err) => {
+        if (err) { throw err; }
+      });
+    });
+  };
+
+  unzipWin = (savePath, unzipTo) => {
+    const { completeInstall } = this.props;
+    const unzipper = new DecompressZip(savePath);
+
+    unzipper.extract({ path: unzipTo });
+
+    // Delete .zip after unzipping
+    unzipper.on('extract', () => {
+      completeInstall();
+      isInstalling = false;
+
+      exec(`DEL ${savePath}`, (error) => {
+        if (error) { throw error; }
+      });
+    });
   };
 
   render() {
